requirements.txt
tensorflow
robosuite
robosuite_models
numpy
scipy
tqdm
pyyaml
tensorboard
hydra-core